pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "winrm-forkbomb-demo"
version = "0.1.0"
description = "Research demo: Ansible WinRM molecule forking behavior and connection exhaustion"
requires-python = ">=3.14,<4.0"

[tool.hatch.build.targets.wheel]
packages = ["tests"]

dependencies = [
  "ansible>=9.0.0",
  "ansible-core>=2.16.0",
  "pywinrm>=0.4.3",
  "pywinrm[credssp]>=0.4.3",
  "pypsrp>=0.8.0",
  "requests-ntlm>=1.2.0",
  "pyspnego>=0.8.0",
  "pyyaml>=6.0",
  "jinja2>=3.1.0",
  "structlog>=24.0.0",
]

[dependency-groups]
test = [
  "molecule>=6.0.3",
  "molecule-plugins>=23.5.0",
  "pytest>=8.0.0",
  "pytest-ansible>=24.0.0",
  "ansible-lint>=6.22.0",
  "yamllint>=1.33.0",
]

[tool.uv]
default-groups = ["test"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "structure: repo structure validation tests (no Windows host needed)",
  "dhall: Dhall output validation tests",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ansible-lint]
profile = "production"
skip_list = ["yaml[line-length]", "name[missing]"]
